[kietnp][modify/create] create flow for province and ward, add new entity Page. Modify all old flow to get rid of MediatR to use Pure CQRS pattern, and remove duplicate PagedList

+++ bmak-ecommerce.Application/Features/Products/DTOs/Catalog/ProductDto.cs
@@ -22,7 +22,7 @@ namespace bmak_ecommerce.Application.Features.Products.DTOs.Catalog
 		// Giá & Kho
 		public decimal Price { get; set; }       // SalePrice
 		public decimal? OriginalPrice { get; set; } // BasePrice
-		public int StockQuantity { get; set; }
+		public float StockQuantity { get; set; }
 
 		// Hình ảnh
 		public string? Thumbnail { get; set; }
